backend/requirements.txt
# Bot Telegram
python-telegram-bot==21.7

# HTTP Requests
requests==2.32.3
aiohttp==3.11.6

# Web Framework
flask==3.1.0
flask-cors==4.0.1

# Utilities
python-dotenv==1.0.0
qrcode==8.0
Pillow==10.4.0

# JSON manipulation
json5==0.9.28

# Date/Time handling
python-dateutil==2.9.0

# Logging
structlog==24.4.0